Refactor.
Gather all state into a structure that can be passed around. Re-write the loops
to be time-based rather than iteration based. Add different speeds and
durations, randomize the new state in the main loop. Add a twinkle mode. Fix
the cylon mode to be wraparound instead of dead ended. Add a reverse mode for
directions.